# Archivo swagger.json/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sales Products API",
    "version": "v1"
  },
  "paths": {
    "/api/Order": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "GetOrders",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDToIEnumerableApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDTo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Order/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OrderDTo": {
        "type": "object",
        "properties": {
          "pedId": {
            "type": "integer",
            "format": "int32"
          },
          "pedUsu": {
            "type": "integer",
            "format": "int32"
          },
          "pedProd": {
            "type": "integer",
            "format": "int32"
          },
          "pedVrUnit": {
            "type": "number",
            "format": "double"
          },
          "pedCant": {
            "type": "number",
            "format": "double"
          },
          "pedSubTot": {
            "type": "number",
            "format": "double"
          },
          "pedIva": {
            "type": "number",
            "format": "double"
          },
          "pedTotal": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OrderDToIEnumerableApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDTo"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {
          "type": "object",
          "additionalProperties": false
        }
      },
...
    }
  }
}